+++ Configuration/TCA/Overrides/sys_template.php
@@ -1,5 +1,7 @@
 <?php
 
+use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
+
 defined('TYPO3') or die('Access denied.');
 
-\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile('seminars', 'Configuration/TypoScript', 'Seminars');
+ExtensionManagementUtility::addStaticFile('seminars', 'Configuration/TypoScript', 'Seminars');